requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
pyodbc==5.0.1
alembic==1.12.1

# Autenticación y HTTP client
httpx==0.25.2

# Procesamiento de archivos
openpyxl==3.1.2
pandas==2.1.3

# Email
aiosmtplib==3.0.1

# Logging y telemetría
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1

# Utilidades
pydantic==2.5.0
